tools/compile_seccomp_policy: Allow permissive default actions
This change provides an escape hatch for overriding the @default action
provided in the .policy file.

.policy files should always have restrictive default actions that
terminate the thread/process upon policy violation. But for debugging
purposes it is super convenient to generate a BPF file that just logs
any violations. So the --default-action flag was added to reconcile both
goals.

Bug: chromium:856315
Test: ./tools/compile_seccomp_policy_unittest.py